// src/services/rest-api/customer-service/model/questions-output.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Customer Service API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.267
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { Choice } from './choice';
// May contain unused imports in some cases
// @ts-ignore
import { Question } from './question';

/**
 * 
 * @export
 * @interface QuestionsOutput
 */
export interface QuestionsOutput {
    /**
     * 
     * @type {Array<Question>}
     * @memberof QuestionsOutput
     */
    'questions'?: Array<Question> | null;
    /**
     * 
     * @type {{ [key: string]: Array<Choice>; }}
     * @memberof QuestionsOutput
     */
    'choices'?: { [key: string]: Array<Choice>; } | null;
}